pyproject.toml
[tool.poetry]
name = "delta-agents"
version = "0.1.0"
description = "Cloud-based sandbox-as-a-service for self-improving LLM agents"
authors = ["DELTA Team <amalampatia1@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/stickfigure3/DELTA"
keywords = ["llm", "agents", "sandbox", "ai", "automation"]

packages = [
    { include = "delta", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
structlog = "^24.1.0"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.0"
redis = "^5.0.0"
argon2-cffi = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
boto3 = "^1.34.0"
twilio = "^8.10.0"
stripe = "^7.0.0"
python-multipart = "^0.0.6"
websockets = "^12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
